cambios en el componente details, y estilos del carrito de la navbar y footer
