+++ client/src/components/Footer/Footer.css
@@ -1,14 +1,14 @@
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400&display=swap');
 
-.container-footer{
-	max-width: 100vw;
-	margin:auto;
-	background: #022E1A;
-	font-family: 'Poppins', sans-serif;
-
+.footer-father{
 	display: flex;
-  	justify-content: center;
-	
+	background: #022E1A;
+	width: 100vw;
+}
+.container-footer{
+	font-family: 'Poppins', sans-serif;
+	width: 100%;
+	margin: 0 auto;
 }
 .row-footer{
 	display: flex;
