+++ client/src/components/ProductDetail/Reviews.css
@@ -1,0 +1,92 @@
+.reviews-div{
+display: flex;
+justify-content: center;
+}
+
+.text-area {
+  width: 550px;
+  border-radius: 5px;
+  margin: 15px 15px;
+  border: solid 1.5px;
+}
+
+.detail-rating input[type="radio"] {
+  display: none;
+}
+
+.star {
+  cursor: pointer;
+  transition: color 200ms;
+}
+
+.text-area {
+  width: 550px;
+  border-radius: 5px;
+  margin: 15px 15px;
+  border: solid 1.5px;
+}
+
+.review {
+  background-color: white;
+  display: grid;
+  border-radius: 5px;
+  margin-bottom: 5px;
+  margin-top: 5px;
+}
+
+.rating {
+  font-size: 16px;
+  font-weight: bold;
+}
+.rating input {
+  border-radius: 4px;
+  margin-left: 10px;
+  margin-right: 5px;
+  height: 20px;
+  width: 70px;
+}
+
+.container-reviews {
+  padding-top: 25px;
+  width: 60%;
+  background-color: #ffffff;
+  margin: 0 auto;
+}
+.container-reviews h1 {
+  color: rgb(0, 0, 0);
+}
+.container-reviews h4 {
+  padding: 0;
+  margin: 5px;
+  text-align: left;
+}
+.container-reviews h5 {
+  font-size: 0.7rem;
+}
+
+.container-reviews element.style {
+  padding: 10px 20px !important;
+}
+
+.container-reviews .MuiPaper-root {
+  background-color: #c4c1c12d;
+  border: 5px solid transparent;
+  border-bottom: 6px solid #17e78c;
+}
+
+.review-button{
+  margin: 0 auto;
+}
+
+.review-button button{
+  padding: 10px 17px;
+  width: 20%;
+  background: #171717;
+  color: #f4f4f4;
+  cursor: pointer;
+}
+
+.review-button button:hover{
+  opacity: 0.9;
+  background-color: #686868;
+}
