+++ client/src/index.css
@@ -13,6 +13,10 @@ a{
   text-decoration: none;
 }
 
+ul li {
+  list-style:none;
+}
+
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
